proyectos_alternativos_1/proyectos_alternativos/chat_embeddings/requirements.txt
sentence-transformers>=2.7.0
scikit-learn
numpy
fastapi
uvicorn[standard]
python-dotenv
